requirements-dev.txt
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==21.4.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
autoflake==1.4
black==21.12b0; python_full_version >= "3.6.2"
cfgv==3.3.1; python_full_version >= "3.6.1"
click==8.0.3; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.6.2" and platform_system == "Windows" and (python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0")
coverage==5.5; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0" and python_version < "4")
distlib==0.3.4; python_full_version >= "3.6.1"
filelock==3.4.2; python_version >= "3.7" and python_full_version >= "3.6.1"
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
flask==2.0.3; python_version >= "3.6"
grpc-example-common @ git+ssh://git@github.com/so1n/grpc-example-common.git@v0.1.5 ; python_version >= "3.8" and python_version < "4.0"
grpcio-testing==1.43.0
grpcio-tools==1.43.0; python_version >= "3.8" and python_version < "4.0"
grpcio==1.43.0; python_version >= "3.8" and python_version < "4.0"
identify==2.4.4; python_full_version >= "3.6.1"
iniconfig==1.1.1; python_version >= "3.6"
isort==5.10.1; python_full_version >= "3.6.1" and python_version < "4.0"
itsdangerous==2.1.0; python_version >= "3.7"
jinja2==3.0.3; python_version >= "3.6"
markupsafe==2.1.0; python_version >= "3.7"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
mypy-extensions==0.4.3; python_full_version >= "3.6.2" and python_version >= "3.5"
mypy-protobuf==2.10; python_version >= "3.6"
mypy==0.910; python_version >= "3.5"
nodeenv==1.6.0; python_full_version >= "3.6.1"
packaging==21.3; python_version >= "3.6"
pathspec==0.9.0; python_full_version >= "3.6.2"
platformdirs==2.4.1; python_version >= "3.7" and python_full_version >= "3.6.2"
pluggy==1.0.0; python_version >= "3.6"
pre-commit==2.17.0; python_full_version >= "3.6.1"
protobuf==3.19.3; python_version >= "3.8" and python_version < "4.0"
py==1.11.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyparsing==3.0.6; python_version >= "3.6"
pytest-grpc==0.8.0
pytest-mock==3.6.1; python_version >= "3.6"
pytest==6.2.5; python_version >= "3.6"
pyyaml==6.0; python_version >= "3.6" and python_full_version >= "3.6.1"
six==1.16.0; python_version >= "3.8" and python_version < "4.0" and python_full_version >= "3.6.1"
toml==0.10.2; python_full_version >= "3.6.1" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
tomli==1.2.3; python_version >= "3.6" and python_full_version >= "3.6.2"
types-futures==3.3.7; python_version >= "3.6"
types-protobuf==3.19.5; python_version >= "3.6"
typing-extensions==4.0.1
virtualenv==20.13.0; python_full_version >= "3.6.1"
werkzeug==2.0.3; python_version >= "3.6"
